fix: adds utils file
